requirements.txt
flask
flask-cors
pymupdf
google-generativeai
gunicorn